Refactor - move message classes to a separate package
