+++ common/src/main/java/com/ably/tracking/common/Ably.kt
@@ -8,6 +8,10 @@ import com.ably.tracking.common.logging.e
 import com.ably.tracking.common.logging.i
 import com.ably.tracking.common.logging.v
 import com.ably.tracking.common.logging.w
+import com.ably.tracking.common.message.getEnhancedLocationUpdate
+import com.ably.tracking.common.message.toJson
+import com.ably.tracking.common.message.toMessage
+import com.ably.tracking.common.message.toMessageJson
 import com.ably.tracking.connection.ConnectionConfiguration
 import com.ably.tracking.logging.LogHandler
 import com.google.gson.Gson
