+++ publishing-sdk/src/main/java/com/ably/tracking/publisher/LocationHistoryModels.kt
@@ -3,10 +3,10 @@ package com.ably.tracking.publisher
 import com.ably.tracking.GeoJsonGeometry
 import com.ably.tracking.GeoJsonMessage
 import com.ably.tracking.GeoJsonProperties
-import com.ably.tracking.common.GEOMETRY_ALT_INDEX
-import com.ably.tracking.common.GEOMETRY_LAT_INDEX
-import com.ably.tracking.common.GEOMETRY_LONG_INDEX
-import com.ably.tracking.common.GeoJsonTypes
+import com.ably.tracking.common.message.GEOMETRY_ALT_INDEX
+import com.ably.tracking.common.message.GEOMETRY_LAT_INDEX
+import com.ably.tracking.common.message.GEOMETRY_LONG_INDEX
+import com.ably.tracking.common.message.GeoJsonTypes
 import com.mapbox.navigation.core.replay.history.ReplayEventBase
 import com.mapbox.navigation.core.replay.history.ReplayEventLocation
 import com.mapbox.navigation.core.replay.history.ReplayEventUpdateLocation
